requirements.txt
Flask>=2.3.3
Flask-Login>=0.6.3
openai>=1.3.7
python-dotenv>=1.0.0
Flask-SQLAlchemy>=3.0.5
Flask-CORS>=4.0.0
Werkzeug>=2.3.7
python-decouple>=3.8
httpx>=0.25.2
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0
pypdf>=3.17.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0 